Added community inactive state check for like/delete post action on 3 screens (updates, user post, community details respectively)

+++ app/src/main/java/com/primapp/ui/post/UpdatesFragment.kt
@@ -12,6 +12,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.primapp.R
 import com.primapp.cache.UserCache
+import com.primapp.constants.CommunityStatusTypes
 import com.primapp.databinding.FragmentUpdatesBinding
 import com.primapp.extensions.showError
 import com.primapp.extensions.showInfo
@@ -207,10 +208,19 @@ class UpdatesFragment : BaseFragment<FragmentUpdatesBinding>() {
                 findNavController().navigate(R.id.videoViewDialog, bundle)
             }
             is LikePost -> {
-                if (item.isLike) {
-                    viewModel.unlikePost(item.communityId, userData!!.id, item.postId)
+                if (item.postData.community.status.equals(CommunityStatusTypes.INACTIVE, true)) {
+                    DialogUtils.showCloseDialog(
+                        requireActivity(),
+                        R.string.inactive_community_action_message,
+                        R.drawable.question_mark
+                    )
+                    return
+                }
+
+                if (item.postData.isLike) {
+                    viewModel.unlikePost(item.postData.community.id, userData!!.id, item.postData.id)
                 } else {
-                    viewModel.likePost(item.communityId, userData!!.id, item.postId)
+                    viewModel.likePost(item.postData.community.id, userData!!.id, item.postData.id)
                 }
             }
 
@@ -219,6 +229,15 @@ class UpdatesFragment : BaseFragment<FragmentUpdatesBinding>() {
             }
 
             is DeletePost -> {
+                if (item.postData.community.status.equals(CommunityStatusTypes.INACTIVE, true)) {
+                    DialogUtils.showCloseDialog(
+                        requireActivity(),
+                        R.string.inactive_community_action_message,
+                        R.drawable.question_mark
+                    )
+                    return
+                }
+
                 DialogUtils.showYesNoDialog(requireActivity(), R.string.delete_post_message, {
                     viewModel.deletePost(item.postData.community.id, userData!!.id, item.postData.id)
                 })
